+++ sct/io/point_target_manager.py
@@ -6,9 +6,6 @@ Point Target and Calibration Sites utilities
 --------------------------------------------
 """
 
-import datetime
-import sqlite3
-from enum import Enum
 from pathlib import Path
 from typing import Union
 
@@ -18,40 +15,11 @@ from arepytools.io import PointSetProduct, read_point_targets_file
 from arepytools.io.io_support import NominalPointTarget
 from arepytools.timing.precisedatetime import PreciseDateTime
 
-from sct import calibration_sites_db, csv_template
-
-SELECT_JOIN_QUERY = """
-SELECT target_name, target_type.type, plate.plate, reference_frame.reference,
-latitude_deg, longitude_deg, altitude_m,
-x_coord_m, y_coord_m, z_coord_m, drift_velocity_x_my, drift_velocity_y_my, drift_velocity_z_my,
-delay_s, measurement_date, validity_start_date, validity_end_date
-FROM XXXX
-INNER JOIN target_type ON target_type.ID = XXXX.target_type
-INNER JOIN plate ON plate.ID = XXXX.plate
-INNER JOIN reference_frame ON reference_frame.ID = XXXX.xyz_reference_frame;
-"""
-
-SELECT_JOIN_QUERY_WITH_TIME = """
-SELECT target_name, target_type.type, plate.plate, reference_frame.reference,
-x_coord_m, y_coord_m, z_coord_m, drift_velocity_x_my, drift_velocity_y_my, drift_velocity_z_my,
-delay_s, measurement_date, validity_start_date, validity_end_date
-FROM XXXX
-WHERE XXXX.validity_start_date <  ?;
-"""
-# INNER JOIN target_type ON target_type.ID = XXXX.target_type
-# INNER JOIN plate ON plate.ID = XXXX.plate
-# INNER JOIN reference_frame ON reference_frame.ID = XXXX.xyz_reference_frame;
-# """
+from sct import csv_template
 
 
 class UnsupportedPointTargetSource(RuntimeError):
     """External point target source provided is not supported"""
-
-
-class SupportedCalibrationSites(Enum):
-    """Supported Calibration Sites inside internal point target database"""
-
-    SURAT_BASIN = "surat_basin"
 
 
 def extract_point_target_data_from_source(source: Union[str, Path]) -> pd.DataFrame:
@@ -93,42 +61,6 @@ def extract_point_target_data_from_source(source: Union[str, Path]) -> pd.DataFr
         raise UnsupportedPointTargetSource(source)
 
     return point_targets_df
-
-
-def query_calibration_sites_db(
-    calibration_site: SupportedCalibrationSites = SupportedCalibrationSites.SURAT_BASIN,
-    acquisition_time: PreciseDateTime = None,
-) -> pd.DataFrame:
-    """Reading calibration sites DB to extract data corresponding to a given site.
-
-    Parameters
-    ----------
-    calibration_site : SupportedCalibrationSites, optional
-        calibration site of interest, by default SupportedCalibrationSites.SURAT_BASIN
-    acquisition_time : PreciseDateTime, optional
-        if provided, query is restricted only to those target valid at the time of acquisition, by default None
-
-    Returns
-    -------
-    pd.DataFrame
-        dataframe containing point target data extracted from db
-    """
-    connection = sqlite3.connect(calibration_sites_db)
-
-    data_df = pd.read_sql_query(
-        SELECT_JOIN_QUERY.replace("XXXX", calibration_site.value),
-        connection,
-        parse_dates=["measurement_date", "validity_start_date", "validity_end_date"],
-    )
-
-    if acquisition_time is not None:
-        acq_date = datetime.datetime(acquisition_time.year, acquisition_time.month, acquisition_time.day_of_the_month)
-        data_df = data_df.query("validity_start_date <= @acq_date & validity_end_date >= @acq_date")
-
-    # substituting all None with Nan
-    data_df = data_df.fillna(value=np.nan)
-
-    return data_df
 
 
 def convert_point_target_binary_to_df(source: Union[str, Path]) -> pd.DataFrame:
@@ -198,7 +130,7 @@ def convert_point_target_file_xml_to_df(source: Union[str, Path]) -> pd.DataFram
 
 
 def convert_df_to_nominal_point_target(data_df: pd.DataFrame) -> dict[str, NominalPointTarget]:
-    """Convert dataframe read from database to dictionary of NominalPointTarget values.
+    """Convert dataframe to dictionary of NominalPointTarget values.
 
     Parameters
     ----------